feat: display warning on Mint VAI UI if user is not Prime
