+++ src/packages/translations/translations/en.json
@@ -837,10 +837,12 @@
       "apy": "APY",
       "connectWallet": "Please connect your wallet to mint VAI",
       "mintFeeLabel": "Mint fee",
+      "primeOnlyWarning": "Only users who hold a Prime token can mint VAI. Learn more about Prime <Link>here</Link>.",
       "rightMaxButtonLabel": "SAFE MAX",
       "submitButtonDisabledLabel": "Enter a valid amount to mint",
       "submitButtonLabel": "Mint",
-      "vaiLimitLabel": "Available VAI limit"
+      "vaiLimitLabel": "Available VAI limit",
+      "walletBalance": "Wallet balance"
     },
     "repayVai": {
       "connectWallet": "Please connect your wallet to repay VAI",
